requirements.txt
pydantic
fastapi
pyyaml